This is the .exe of my game
I originally started this project in the wrong repository, but then when I tried to fix that mistake I ended up cloning this repository back into the file I was editing. It still has all the functionalities I put in it, that is to say that it can still roll dice and display them, lock certain dice, it does keep track of the number of rolls in a turn by simply having each button disable itself after it's not supposed to be used anymore, then all the buttons reset when the user calculates their score. It also doesn't have the function that detects a full house. I didn't quite get that working. I also don't know how to edit this any more, because I can't find the actual Visual Studio file it was originally in...